+++ src/ckanapi_harvesters/__init__.py
@@ -42,7 +42,8 @@ from . import builder
 from . import reports
 
 # usage shortcuts
-from .auxiliary import CkanMap, RequestType, CkanField, CkanState, CkanProgressCallback, CkanCallbackLevel, CkanProgressBarType
+from .auxiliary import (CkanMap, RequestType, CkanField, CkanState, CkanProgressCallback, CkanCallbackLevel,
+                        CkanProgressBarType, CkanProgressCallbackEmpty)
 from .policies import CkanPackageDataFormatPolicy
 from .reports import CkanAdminReport
 from .ckan_api import CkanApi, CkanApiParams, CKAN_API_VERSION