+++ src/ckanapi_harvesters/builder/builder_resource_datastore_multi_harvester.py
@@ -7,6 +7,7 @@ This concrete implementation is linked to the file system.
 from typing import List, Collection, Any, Tuple, Generator, Union
 from collections import OrderedDict
 from warnings import warn
+import shlex
 
 import pandas as pd
 
@@ -52,8 +53,9 @@ class BuilderDataStoreHarvester(BuilderDataStoreFolder):
         self._update_metadata(None)
 
     def initialize_extra_options_string(self, extra_options_string:str, base_dir:str) -> None:
-        super().initialize_extra_options_string(extra_options_string, base_dir=base_dir)
-        self.harvester = init_table_harvester_from_options_string(self.options_string, file_url_attr=self.file_url_attr, base_dir=base_dir)
+        self.harvester, extra_extra_args = init_table_harvester_from_options_string(extra_options_string, file_url_attr=self.file_url_attr, base_dir=base_dir)
+        extra_extra_options_string = shlex.join(extra_extra_args)
+        super().initialize_extra_options_string(extra_extra_options_string, base_dir=base_dir)
 
     def init_options_from_ckan(self, ckan:CkanApi, *, base_dir:str=None) -> None:
         self.harvester.update_from_ckan(ckan)