+++ src/ckanapi_harvesters/builder/builder_resource_multi_datastore.py
@@ -195,12 +195,13 @@ class BuilderMultiDataStore(BuilderMultiFile, BuilderDataStoreABC):
             return df_local
 
     def upload_file_chunk(self, ckan:CkanApi, package_id:str, file_chunk:FileChunkDataFrame, *,
-                          reupload:bool=False, override_ckan:bool=False, cancel_if_present:bool=True) -> CkanResourceInfo:
+                          reupload:bool=False, override_ckan:bool=False, cancel_if_present:bool=True,
+                          inhibit_datastore_patch_indexes:bool=False) -> CkanResourceInfo:
         file_path = file_chunk.file_path
         ds_builder, file_dir = self._data_store_builder_of_file(file_path=file_path)
         if file_chunk.is_first_chunk:
             return ds_builder.patch_request(ckan=ckan, package_id=package_id, reupload=reupload,
-                                            resources_base_dir=file_dir)
+                                            resources_base_dir=file_dir, inhibit_datastore_patch_indexes=inhibit_datastore_patch_indexes)
         else:
             ds_builder.upsert_request_df(ckan, file_chunk.df, file_name=file_path, total_lines_read=self.read_line_counter)
             return ckan.map.get_resource_info(resource_name=ds_builder.name, package_name=package_id)